gripai-reports/requirements.txt
anthropic>=0.40.0
supabase>=2.0.0
resend>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.0
